requirements.txt
django >=1.8, <1.9
feedparser >=5.2, <5.3
pyhamcrest >=1.8, <1.9
python-dateutil >=2.4, <2.5
mock >=1.0, <1.1